requirements.txt
django-s3-storage==0.12.4
mysqlclient==1.3.14
